backend/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
SQLAlchemy
psycopg2-binary
jwt
PyJWT
pydantic[email]
passlib
python-multipart
emails
python-jose
ipython
langchain
langgraph
langchain-google-genai
google-generativeai
beautifulsoup4
langchain-tavily
bcrypt=4.0.1 